requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.3
django-cors-headers==2.4.0
django-import-export==1.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.0
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.4.0
openpyxl==2.5.12
Pillow==5.3.0
pytz==2018.7
PyYAML==3.13
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.1.0
xlwt==1.3.0
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2==2.7.5
psycopg2-binary==2.7.5